+++ src/DrDotnet.Profilers/profilers/gc_survivors_profiler.rs
@@ -5,7 +5,6 @@ use std::hash::BuildHasherDefault;
 use std::ops::AddAssign;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::thread;
-use std::time::Instant;
 use thousands::{digits, Separable, SeparatorPolicy};
 
 use crate::api::*;
@@ -20,9 +19,9 @@ pub struct GCSurvivorsProfiler {
     name_resolver: CachedNameResolver,
     clr_profiler_info: ClrProfilerInfo,
     session_info: SessionInfo,
-    object_to_referencers: HashMap<ObjectID, Option<Vec<ObjectID>>, BuildHasherDefault<SimpleHasher>>,
-    is_triggered_gc: AtomicBool,
-    gc_start_time: Option<Instant>,
+    root_objects: Vec<ObjectID>,
+    root_kinds: HashMap<ObjectID, COR_PRF_GC_ROOT_KIND, BuildHasherDefault<SimpleHasher>>,
+    is_relevant_gc: AtomicBool,
 }
 
 #[derive(Clone, Default, Debug)]
@@ -51,8 +50,9 @@ impl Display for References {
         } else {
             "???".to_string()
         };
+        let nb_objects_str = nb_objects.separate_by_policy(policy);
 
-        write!(f, "({total_size_str} bytes) - {nb_objects}")
+        write!(f, "{nb_objects_str} / {total_size_str} B")
     }
 }
 
@@ -61,42 +61,34 @@ impl Profiler for GCSurvivorsProfiler {
 
     fn profiler_info() -> ProfilerInfo {
         return ProfilerInfo {
-            uuid: "805A308B-061C-47F3-9B30-F785C3186E86".to_owned(),
+            uuid: "805A307B-061C-47F3-9B30-F795C3186E86".to_owned(),
             name: "List GC survivors".to_owned(),
-            description: "Triggers a full block garbage collection, and then lists surviving references as a dependency tree.\n\n*⚠️ Experimental. This profiler can be very slow due to the huge amount of data it has to deal with.*".to_owned(),
+            description: "Perform a full blocking garbage collection and list the objects that survived it, grouped by type and sorted by size or count of retained objects. This helps to identify memory leaks, understand the retention paths of objects in memory and track down ephemeral objects that usually lengthen the time to perform a garbage collection.".to_owned(),
             parameters: vec![
-                ProfilerParameter {
-                    name: "Sort by size".to_owned(),
-                    key: "sort_by_size".to_owned(),
-                    description: "If true, sort the results by inclusive size (bytes). Otherwise, sort by inclusive instances count.".to_owned(),
-                    type_: ParameterType::BOOLEAN.into(),
-                    value: "true".to_owned(),
-                    ..std::default::Default::default()
-                },
-                ProfilerParameter {
-                    name: "Sort multi-threaded".to_owned(),
-                    key: "sort_multithreaded".to_owned(),
-                    description: "If true, sort the results with a multi-threaded methode. Otherwise, sort with an iterative method.".to_owned(),
-                    type_: ParameterType::BOOLEAN.into(),
-                    value: "false".to_owned(),
-                    ..std::default::Default::default()
-                },
-                ProfilerParameter {
-                    name: "Maximum types to display".to_owned(),
-                    key: "max_types_display".to_owned(),
-                    description: "The maximum number of types to display in the report".to_owned(),
-                    type_: ParameterType::INT.into(),
-                    value: "1000".to_owned(),
-                    ..std::default::Default::default()
-                },
-                ProfilerParameter {
-                    name: "Maximum depth".to_owned(),
-                    key: "max_retention_depth".to_owned(),
-                    description: "The maximum depth while drilling through retention paths".to_owned(),
-                    type_: ParameterType::INT.into(),
-                    value: "4".to_owned(),
-                    ..std::default::Default::default()
-                },
+                ProfilerParameter::define(
+                    "Sort by size",
+                    "sort_by_size",
+                    false,
+                    "If true, sort the results by inclusive size (bytes). Otherwise, sort by (inclusive) count of retained object. Disable this option to improve performance.",
+                ),
+                ProfilerParameter::define(
+                    "Retained references threshold",
+                    "retained_references_threshold",
+                    100,
+                    "Threshold of number of retained references by a node to ignore it. This helps to reduce the tree size, improving performance and readability.",
+                ),
+                ProfilerParameter::define(
+                    "Retained bytes threshold",
+                    "retained_bytes_threshold",
+                    10000,
+                    "Threshold of number of retained bytes by a node to ignore it. This helps to reduce the tree size, improving performance and readability.",
+                ),
+                ProfilerParameter::define(
+                    "Maximum depth",
+                    "max_depth",
+                    4,
+                    "The maximum depth while drilling through retention paths. This helps to reduce the tree size, improving performance and readability.",
+                ),
             ],
             ..std::default::Default::default()
         };
@@ -104,139 +96,121 @@ impl Profiler for GCSurvivorsProfiler {
 }
 
 impl GCSurvivorsProfiler {
-    pub fn append_references(&self, info: &ClrProfilerInfo, object_id: ffi::ObjectID, max_depth: usize) -> Vec<Vec<ClassID>> {
-        let mut branches: Vec<Vec<ClassID>> = Vec::new();
+    fn gather_references_recursive(
+        info: &ClrProfilerInfo,
+        node: &mut TreeNode<ClassID, References>,
+        depth: usize,
+        max_depth: usize,
+        retained_references_threshold: usize,
+        retained_bytes_threshold: usize,
+    ) {
+        if depth > max_depth {
+            // Keep this node as-is, don't drill further
+            return;
+        }
 
-        let mut branch: Vec<ClassID> = vec![]; // A branch we update live and copy every time we reached the end of a path
-        let mut stack: Vec<(ObjectID, usize)> = vec![(object_id, 0)]; // A stack to iterate without recursions
+        if let Some(references) = &node.value {
+            let reference_object_ids = Vec::<ObjectID>::new();
 
-        while !stack.is_empty() {
-            let current_id: (ObjectID, usize) = stack.pop().unwrap();
-
-            // Trim branch until it has the proper depth
-            while branch.len() > current_id.1 {
-                branch.pop();
-            }
-
-            match info.get_class_from_object(current_id.0) {
-                Ok(class_id) => branch.push(class_id),
-                Err(_) => {
-                    error!("Impossible to get class ID for object ID {}", current_id.0);
+            // For each object instance, enumerate its references
+            for &object_id in references.0.keys() {
+                if object_id == 0 {
                     continue;
                 }
+                // We must pass this data as a pointer for callback to mutate it with actual object references ids
+                let references_ptr_c = &reference_object_ids as *const Vec<ObjectID> as *mut std::ffi::c_void;
+                let _ = info.enumerate_object_references(object_id, crate::utils::enum_references_callback, references_ptr_c);
             }
 
-            match self.object_to_referencers.get(&current_id.0) {
-                Some(referencers) => match referencers {
-                    Some(referencers) => {
-                        if current_id.1 < max_depth {
-                            // Push new referencers if we are within our allowed depth
-                            for i in 0..referencers.len() {
-                                stack.push((referencers[i], current_id.1 + 1));
-                            }
-                        } else {
-                            // If max depth is reached, we push a 0 terminated branch to indicate that branch is truncated
-                            let mut deep_branch = branch.clone();
-                            deep_branch.push(0);
-                            branches.push(deep_branch);
-                        }
-                    }
-                    None => {
-                        // We reached the end of a path, copy the branch and add it to our branches
-                        branches.push(branch.clone());
-                    }
-                },
-                None => {
-                    // We reached the end of a path, copy the branch and add it to our branches
-                    branches.push(branch.clone());
-                }
-            }
+            Self::build_tree_nodes(
+                info,
+                &reference_object_ids,
+                node,
+                depth,
+                max_depth,
+                retained_references_threshold,
+                retained_bytes_threshold,
+            );
         }
-
-        // Example:
-        // M
-        // ├─ A
-        // │  └─ D
-        // ├─ B
-        // └─ C
-        //    ├─ F
-        //    └─ G
-        // 0  1  2
-        // Iteration 1: We pop M, which is added to the branch. A, B and C are pushed to the stack
-        // Iteration 2: We pop C, which is added to the branch. F and G are pushed to the stack
-        // Iteration 3: We pop G, which is added to the branch. There are no child, so the branch MCG is completed
-        // Iteration 4: We pop F. F depth is 2, but the branch is currently of len 3, so G is removed from the branch. F is then added to the branch. There are no child, so the branch MCF is completed
-        // Iteration 5: We pop B. B depth is 1, but the branch is currently of len 3, so C and F are removed from the branch. B is then added to the branch. There are no child, so the branch MB is completed
-        // Iteration 6: We pop A. A depth is 1, but the branch is currently of len 2, so B is removed from the branch. A is then added to the branch
-        // Iteration 7: We pop D. which is added to the branch. There are no child, so the branch MAD is completed
-        // The stack is now empty, we're done :)
-
-        return branches;
     }
 
-    // Post-process tracked persisting references
-    fn build_sequences(&mut self) -> HashMap<Vec<usize>, References> {
-        info!(
-            "Building graph of surviving references... {} objects to process",
-            self.object_to_referencers.len()
-        );
+    fn build_tree_nodes(
+        clr: &ClrProfilerInfo,
+        reference_object_ids: &Vec<ObjectID>,
+        node: &mut TreeNode<usize, References>,
+        depth: usize,
+        max_retention_depth: usize,
+        retained_references_threshold: usize,
+        retained_bytes_threshold: usize
+    ) {
+        let mut map: HashMap<ClassID, References> = HashMap::new();
 
-        let now = std::time::Instant::now();
-
-        let mut sequences: HashMap<Vec<ClassID>, References> = HashMap::new();
-
-        let info = self.clr();
-
-        let max_retention_depth = self.session_info().get_parameter::<usize>("max_retention_depth").unwrap();
-
-        for object in self.object_to_referencers.iter() {
-            let object_id: ObjectID = object.0.clone();
-
-            if !Self::is_gen_2(info, object_id) {
+        // Fill a map of class_id -> references
+        // This will allow us to group references by class and ease the creation of child nodes in the tree
+        for &object_id in reference_object_ids.iter() {
+            if object_id == 0 {
                 continue;
             }
-
-            let size = info.get_object_size_2(object_id).unwrap_or(0);
-
-            for branch in self.append_references(info, object_id, max_retention_depth) {
-                sequences
-                    .entry(branch)
-                    .and_modify(|referencers| {
-                        referencers.0.insert(object_id.clone(), size);
-                    })
-                    .or_insert_with(|| {
-                        let mut map = HashMap::default();
-                        map.insert(object_id.clone(), size);
-                        References(map)
-                    });
-            }
+            let refs = map.entry(clr.get_class_from_object(object_id).unwrap()).or_insert(References::default());
+            let size = clr.get_object_size_2(object_id).unwrap_or(0);
+            refs.0.insert(object_id, size);
         }
 
-        // Free some memory
-        self.object_to_referencers = HashMap::default();
+        // For each class_id, create a child node and fill it with references
+        for (class_id, refs) in map {
+            let mut child_node = TreeNode::new(class_id);
+            child_node.value = Some(refs);
+            Self::gather_references_recursive(
+                clr,
+                &mut child_node,
+                depth + 1,
+                max_retention_depth,
+                retained_references_threshold,
+                retained_bytes_threshold,
+            );
 
-        info!("Graph built in {} ms", 0.001 * now.elapsed().as_micros() as f64);
-
-        sequences
+            // Discard the branch if it doesn't meet the thresholds
+            // This will save memory, lower CPU usage when sorting the tree and lighten the final report
+            let inclusive_values = child_node.get_inclusive_value().0;
+            let mut discard_branch = inclusive_values.len() < retained_references_threshold;
+            discard_branch |= inclusive_values.values().sum::<usize>() < retained_bytes_threshold;
+            if !discard_branch {
+                node.children.push(child_node);
+            }
+        }
     }
 
-    fn is_gen_2(info: &ClrProfilerInfo, object_id: usize) -> bool {
-        info.get_object_generation(object_id)
-            .map_or(false, |gen_info| gen_info.generation == ffi::COR_PRF_GC_GENERATION::COR_PRF_GC_GEN_2)
-    }
-
-    fn build_tree(&self, sequences: &mut HashMap<Vec<usize>, References>) -> TreeNode<usize, References> {
-        info!("Building tree");
+    fn build_tree(&mut self) -> TreeNode<ClassID, References> {
+        info!("Building tree of surviving references from {} roots...", self.root_objects.len());
 
         let now = std::time::Instant::now();
 
-        let mut tree = TreeNode::build_from_sequences(&sequences, 0);
+        let mut tree = TreeNode::new(0);
 
-        sequences.clear();
+        let max_retention_depth = self.session_info().get_parameter::<usize>("max_depth").unwrap();
+        let retained_references_threshold = self.session_info().get_parameter::<usize>("retained_references_threshold").unwrap();
+        let retained_bytes_threshold = self.session_info().get_parameter::<usize>("retained_bytes_threshold").unwrap();
 
-        info!("Tree built in {} ms", 0.001 * now.elapsed().as_micros() as f64);
+        info!("Build root node refs...");
+        let clr = self.clr();
 
-        info!("Sorting tree");
+        Self::build_tree_nodes(
+            clr,
+            &self.root_objects,
+            &mut tree,
+            0,
+            max_retention_depth,
+            retained_references_threshold,
+            retained_bytes_threshold,
+        );
+
+        info!("Tree built in {} ms", now.elapsed().as_millis());
+
+        tree
+    }
+
+    fn sort_tree(&self, tree: &mut TreeNode<usize, References>) {
+        info!("Sorting tree...");
 
         let now = std::time::Instant::now();
 
@@ -256,21 +230,17 @@ impl GCSurvivorsProfiler {
             }
         };
 
+        // Start by sorting the tree "roots" (only the first level of childrens)
+        tree.children.sort_by(compare);
+
         // Then sort the whole tree (all levels of childrens)
-        let sort_multithreaded = self.session_info().get_parameter::<bool>("sort_multithreaded").unwrap();
-        if sort_multithreaded {
-            tree.sort_by_multithreaded(compare);
-        } else {
-            tree.sort_by_iterative(compare);
-        }
+        tree.sort_by_iterative(compare);
 
-        info!("Tree sorted in {} ms", 0.001 * now.elapsed().as_micros() as f64);
-
-        return tree;
+        info!("Tree sorted in {} ms", now.elapsed().as_millis());
     }
 
     fn write_report(&mut self, tree: TreeNode<usize, References>) -> Result<(), HRESULT> {
-        info!("Building report");
+        info!("Writing report...");
 
         let now = std::time::Instant::now();
 
@@ -280,21 +250,46 @@ impl GCSurvivorsProfiler {
         let mut report = self.session_info.create_report("summary.html".to_owned());
 
         report.write_line(format!("<h2>GC Survivors Report</h2>"));
-        report.write_line(format!("<h3>{nb_objects} surviving objects of {nb_classes} classes</h3>"));
+        report.write_line(format!("The GC survivors report displays a tree of objects that survived the last garbage collection. The first level of the tree represents the roots of the objects grouped by typed. Each node of the tree represents a class of objects that are retained by its parent node, and so on until there are no more references to follow or maximum depth is reached. The tree is sorted by the number of retained objects or bytes, depending on the selected parameter."));
+
+        // Quick legend
+        report.write_line(format!(" \
+            <details> \
+                <summary> \
+                    <code>MyRootObject</code> \
+                    <div class=\"chip\"><span>retained objects including all children / retained bytes including all children</span><i class=\"material-icons\">radio_button_checked</i></div> \
+                    <div class=\"chip\"><span>retained objects / retained bytes</span><i class=\"material-icons\">radio_button_unchecked</i></div> \
+                    <div class=\"chip\"><span>stack</span><i class=\"material-icons\">segment</i></div> \
+                    <div class=\"chip\"><span>handle</span><i class=\"material-icons\">point_scan</i></div> \
+                    <div class=\"chip\"><span>finalizer</span><i class=\"material-icons\">auto_delete</i></div> \
+                    <div class=\"chip\"><span>other</span><i class=\"material-icons\">help</i></div> \
+                </summary> \
+                <ul><li> \
+                    <code>MySurvivingObject</code> \
+                    <div class=\"chip\"><span>retained objects including all children / retained bytes including all children</span><i class=\"material-icons\">radio_button_checked</i></div> \
+                    <div class=\"chip\"><span>retained objects / retained bytes</span><i class=\"material-icons\">radio_button_unchecked</i></div> \
+                </li></ul> \
+            </details>"));
+        
+        report.write_line(format!("<h3>Retention Tree</h3>"));
+        report.write_line(format!("<h4>{nb_objects} surviving objects of {nb_classes} classes</h4>"));
 
         for tree_node in tree.children {
-            self.print_html(&tree_node, &mut report);
+            self.print_html(&tree_node, 0, &mut report);
         }
 
-        self.session_info.finish();
-
-        info!("Report written in {} ms", 0.001 * now.elapsed().as_micros() as f64);
+        info!("Report written in {} ms", now.elapsed().as_millis());
 
         Ok(())
     }
 
-    fn print_html(&self, tree: &TreeNode<ClassID, References>, report: &mut Report) {
-        let refs = &tree.get_inclusive_value();
+    fn print_html(&self, tree: &TreeNode<ClassID, References>, depth: usize, report: &mut Report) {
+
+        let references_exlusive = match &tree.value {
+            None => &References::default(),
+            Some(refs) => refs
+        };
+        let references_inclusive = &tree.get_inclusive_value();
 
         if tree.key == 0 {
             report.write_line(format!("Path truncated because of depth limit reached"));
@@ -306,55 +301,44 @@ impl GCSurvivorsProfiler {
 
         let has_children = tree.children.len() > 0;
 
+        let mut line: String = format!("<code>{escaped_class_name}</code> \
+            <div class=\"chip\"><span>{references_inclusive}</span><i class=\"material-icons\">radio_button_checked</i></div> \
+            <div class=\"chip\"><span>{references_exlusive}</span><i class=\"material-icons\">radio_button_unchecked</i></div>");
+
+        if depth == 0 {
+            // Find index of first item with id 42
+            let mut count_per_kind = HashMap::new();
+            for (object_id, size) in &tree.value.as_ref().unwrap().0 {
+                let kind = self.root_kinds.get(&object_id).unwrap();
+                let count = count_per_kind.entry(kind).or_insert(0);
+                *count += 1;
+            }
+            for (kind, count) in count_per_kind {
+                let kind_icon = match kind {
+                    ffi::COR_PRF_GC_ROOT_KIND::COR_PRF_GC_ROOT_STACK => "segment",
+                    ffi::COR_PRF_GC_ROOT_KIND::COR_PRF_GC_ROOT_FINALIZER => "auto_delete",
+                    ffi::COR_PRF_GC_ROOT_KIND::COR_PRF_GC_ROOT_HANDLE => "point_scan",
+                    ffi::COR_PRF_GC_ROOT_KIND::COR_PRF_GC_ROOT_OTHER => "help"
+                };
+                line = format!("{line}<div class=\"chip\"><span>{count}</span><i class=\"material-icons\">{kind_icon}</i></div>");
+            }
+        }
+
         if has_children {
-            report.write_line(format!("<details><summary><span>{refs}</span><code>{escaped_class_name}</code></summary>"));
+            report.write_line(format!("<details><summary>{line}</summary>"));
             report.write_line(format!("<ul>"));
             for child in &tree.children {
-                self.print_html(child, report);
+                self.print_html(child, depth + 1, report);
             }
             report.write_line(format!("</ul>"));
             report.write_line(format!("</details>"));
         } else {
-            report.write_line(format!("<li><span>{refs}</span><code>{escaped_class_name}</code></li>"));
+            report.write_line(format!("<li>{line}</li>"));
         }
     }
 }
 
-impl CorProfilerCallback for GCSurvivorsProfiler {
-    fn object_references(&mut self, object_id: ffi::ObjectID, class_id: ffi::ClassID, object_ref_ids: &[ffi::ObjectID]) -> Result<(), HRESULT> {
-        if !self.is_triggered_gc.load(Ordering::Relaxed) {
-            error!("Early return of object_references because GC wasn't forced yet");
-            // Early return if we received an event before the forced GC started
-            return Ok(());
-        }
-
-        // If an object has no references and is not gen 2, we can discard it, because it means it will never reference any
-        // other gen 2 object, which is actually what we are looking for
-        if object_ref_ids.is_empty() && !Self::is_gen_2(self.clr(), object_id) {
-            return Ok(());
-        }
-
-        // Create dependency tree, but from object to referencers, instead of object to its references.
-        // This is usefull for being able to browse from any object back to its roots.
-        for object_ref_id in object_ref_ids {
-            self.object_to_referencers
-                .entry(*object_ref_id)
-                .and_modify(|e| {
-                    if let Some(vec) = e.as_mut() {
-                        vec.push(object_id);
-                    } else {
-                        *e = Some(vec![object_id]);
-                    }
-                })
-                .or_insert_with(|| Some(vec![object_id]));
-        }
-
-        // Also add this object, with no referencers, just in case this object isn't referenced
-        self.object_to_referencers.entry(object_id).or_insert(None);
-
-        Ok(())
-    }
-}
+impl CorProfilerCallback for GCSurvivorsProfiler {}
 
 impl CorProfilerCallback2 for GCSurvivorsProfiler {
     fn garbage_collection_started(&mut self, generation_collected: &[ffi::BOOL], reason: ffi::COR_PRF_GC_REASON) -> Result<(), HRESULT> {
@@ -362,16 +346,10 @@ impl CorProfilerCallback2 for GCSurvivorsProfiler {
 
         info!("garbage_collection_started on gen {} for reason {:?}", gen, reason);
 
-        // if gen == 1 {
-        //     self.is_triggered_gc.store(true, Ordering::Relaxed);
-        // }
-
         if reason == ffi::COR_PRF_GC_REASON::COR_PRF_GC_INDUCED {
-            info!("induced gc!");
-            self.is_triggered_gc.store(true, Ordering::Relaxed);
+            debug!("induced gc!");
+            self.is_relevant_gc.store(true, Ordering::Relaxed);
         }
-
-        self.gc_start_time = Some(std::time::Instant::now());
 
         Ok(())
     }
@@ -379,32 +357,55 @@ impl CorProfilerCallback2 for GCSurvivorsProfiler {
     fn garbage_collection_finished(&mut self) -> Result<(), HRESULT> {
         info!("garbage_collection_finished");
 
-        if !self.is_triggered_gc.load(Ordering::Relaxed) {
+        if !self.is_relevant_gc.load(Ordering::Relaxed) {
             error!("Early return of garbage_collection_finished because GC wasn't forced yet");
             // Early return if we received an event before the forced GC started
             return Ok(());
         }
 
-        info!(
-            "Garbage collection done in {} ms",
-            0.001 * self.gc_start_time.unwrap().elapsed().as_micros() as f64
-        );
-
         // Disable profiling to free some resources
-        match self.clr().set_event_mask(ffi::COR_PRF_MONITOR::COR_PRF_MONITOR_NONE) {
+        match self
+            .clr()
+            .set_event_mask_2(ffi::COR_PRF_MONITOR::COR_PRF_MONITOR_NONE, ffi::COR_PRF_HIGH_MONITOR::COR_PRF_HIGH_MONITOR_NONE)
+        {
             Ok(_) => (),
             Err(hresult) => error!("Error setting event mask: {:?}", hresult),
         }
 
-        info!("Deep size of object references: {} bytes", self.object_to_referencers.deep_size_of());
+        info!("Deep size of roots: {} bytes", self.root_objects.deep_size_of());
 
-        let mut sequences = self.build_sequences();
-        let tree = self.build_tree(&mut sequences);
+        let mut tree = self.build_tree();
+        self.sort_tree(&mut tree);
         let _ = self.write_report(tree);
 
         // We're done, we can detach :)
         let profiler_info = self.clr().clone();
         profiler_info.request_profiler_detach(3000).ok();
+
+        Ok(())
+    }
+
+    fn root_references_2(
+        &mut self,
+        root_ref_ids: &[ObjectID],
+        root_kinds: &[COR_PRF_GC_ROOT_KIND],
+        root_flags: &[COR_PRF_GC_ROOT_FLAGS],
+        root_ids: &[UINT_PTR], // TODO: Maybe this should be a single array of some struct kind.
+    ) -> Result<(), HRESULT> {
+        if !self.is_relevant_gc.load(Ordering::Relaxed) {
+            warn!("Early return of garbage_collection_finished because GC wasn't forced yet");
+            // Early return if we received an event before the forced GC started
+            return Ok(());
+        }
+
+        for i in 0..root_ref_ids.len() {
+            let root_id = root_ref_ids[i];
+            let root_kind = root_kinds[i];
+            //let root_flag = root_flags[i];
+
+            self.root_objects.push(root_id);
+            self.root_kinds.insert(root_id, root_kind);
+        }
 
         Ok(())
     }
@@ -440,6 +441,11 @@ impl CorProfilerCallback3 for GCSurvivorsProfiler {
         // Security timeout
         detach_after_duration::<GCSurvivorsProfiler>(&self, 320);
 
+        Ok(())
+    }
+
+    fn profiler_detach_succeeded(&mut self) -> Result<(), ffi::HRESULT> {
+        self.session_info.finish();
         Ok(())
     }
 }
